requirements1.txt
Flask
flask-cors
requests
python-dotenv
gunicorn